Extracting client test properties
