+++ service/src/main/java/tds/config/web/endpoints/ConfigController.java
@@ -3,13 +3,16 @@ package tds.config.web.endpoints;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
 import tds.common.web.exceptions.NotFoundException;
 import tds.config.ClientSystemFlag;
-import tds.config.ClientTestProperty;
 import tds.config.services.ConfigService;
 import tds.config.web.resources.ClientSystemFlagResource;
-import tds.config.web.resources.ClientTestPropertyResource;
 
 
 @RestController
@@ -29,14 +32,5 @@ public class ConfigController {
                 .orElseThrow(() -> new NotFoundException("Could not find ClientSystemFlag for client name %s and audit object %s", clientName, auditObject));
 
         return ResponseEntity.ok(new ClientSystemFlagResource(clientSystemFlag));
-    }
-
-    @RequestMapping(value = "/client-test-properties/{clientName}/{assessmentId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    @ResponseBody
-    public ResponseEntity<ClientTestPropertyResource> getClientTestProperty(@PathVariable final String clientName, @PathVariable final String assessmentId) {
-        final ClientTestProperty clientTestProperty = configService.findClientTestProperty(clientName, assessmentId)
-                .orElseThrow(() -> new NotFoundException("Could not find ClientTestProperty for client name %s and assessment id %s", clientName, assessmentId));
-
-        return ResponseEntity.ok(new ClientTestPropertyResource(clientTestProperty));
     }
 }
